actions/requirements.txt
langchain
openai
python-dotenv
faiss-cpu
dateparser
tiktoken
unstructured
boto3
pyodbc